Splitting out network-related stuff
I want the core library to be buildable without the ASIO dependency, or any assumptions on the host application's network stack. If you have an existing network stack capable of querying DNS records, you should be able to use that, rather than having to pull in ASIO for one thing.